Further improvements of unlock notification
