+++ translations/base-en.yaml
@@ -424,6 +424,11 @@ storyRewards:
         desc: >-
             This level gave you no reward, but the next one will! <br><br> PS: Better don't destroy your existing factory - You need <strong>all</strong> those shapes later again to <strong>unlock upgrades</strong>!
 
+    no_reward_freeplay:
+        title: Next level
+        desc: >-
+            Congratulations! By the way, more content is planned for the standalone!
+
 settings:
     title: Settings
     categories:
